Updated Regex to match new format
The format for the highlights package was different for stage 5.  I *think* this new regex handles it, but haven't tested it.  I also don't know golang, so there might be a better way of handling the difference.  This was more meant to point out the change rather than be the actual fix.